@@ -1,6 +1,6 @@
 {
   "name": "tokyo-free-white-react-admin-dashboard",
-  "version": "1.0.0",
+  "version": "1.1.0",
   "title": "Tokyo Free White React Admin Dashboard",
   "description": "This free and open source admin dashboard template is built for React and it’s bootstrapped from Facebook’s create-react-app. All NPM dependencies are up to date and it contains multiple fully customized components based on the popular frontend components framework, Material-UI.",
   "author": {
@@ -9,34 +9,34 @@
   },
   "private": false,
   "dependencies": {
-    "@emotion/react": "11.4.0",
+    "@emotion/react": "11.4.1",
     "@emotion/styled": "11.3.0",
-    "@material-ui/core": "5.0.0-beta.2",
-    "@material-ui/icons": "5.0.0-beta.1",
-    "@material-ui/lab": "5.0.0-alpha.41",
-    "@material-ui/styles": "5.0.0-beta.2",
+    "@mui/material": "5.0.3",
+    "@mui/icons-material": "5.0.3",
+    "@mui/lab": "5.0.0-alpha.50",
+    "@mui/styles": "5.0.1",
     "@types/nprogress": "0.2.0",
-    "@types/react": "17.0.15",
+    "@types/react": "17.0.25",
     "@types/react-dom": "17.0.9",
-    "@types/react-router-dom": "5.1.8",
+    "@types/react-router-dom": "5.3.0",
     "chart.js": "2.9.4",
     "clsx": "1.1.1",
-    "date-fns": "2.23.0",
-    "history": "5.0.0-beta.9",
+    "date-fns": "2.25.0",
+    "history": "5.0.1",
     "nprogress": "0.2.0",
     "numeral": "2.0.6",
     "prop-types": "15.7.2",
     "react": "17.0.2",
-    "react-chartjs-2": "2.11.1",
     "react-custom-scrollbars-2": "4.4.0",
     "react-dom": "17.0.2",
-    "react-helmet-async": "1.0.9",
+    "react-helmet-async": "1.1.2",
     "react-router": "6.0.0-beta.0",
     "react-router-dom": "6.0.0-beta.0",
     "react-scripts": "4.0.3",
     "stylis": "4.0.10",
     "stylis-plugin-rtl": "2.1.0",
-    "web-vitals": "2.1.0"
+    "react-chartjs-2": "2.11.1",
+    "web-vitals": "2.1.2"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -62,21 +62,24 @@
     ]
   },
   "devDependencies": {
-    "@types/numeral": "2.0.1",
-    "@babel/eslint-parser": "7.14.7",
-    "@typescript-eslint/eslint-plugin": "4.28.5",
-    "@typescript-eslint/parser": "4.28.5",
-    "eslint": "7.31.0",
+    "@babel/eslint-parser": "7.15.8",
+    "@types/numeral": "2.0.2",
+    "@types/react": "17.0.29",
+    "@types/react-dom": "17.0.9",
+    "@types/react-router-dom": "5.3.1",
+    "@typescript-eslint/eslint-plugin": "5.0.0",
+    "@typescript-eslint/parser": "5.0.0",
+    "eslint": "7.32.0",
     "eslint-config-airbnb": "18.2.1",
-    "eslint-config-airbnb-typescript": "12.3.1",
+    "eslint-config-airbnb-typescript": "14.0.1",
     "eslint-config-prettier": "8.3.0",
-    "eslint-plugin-import": "2.23.4",
+    "eslint-plugin-import": "2.25.2",
     "eslint-plugin-jsx-a11y": "6.4.1",
-    "eslint-plugin-prettier": "3.4.0",
-    "eslint-plugin-react": "7.24.0",
+    "eslint-plugin-prettier": "4.0.0",
+    "eslint-plugin-react": "7.26.1",
     "eslint-plugin-react-hooks": "4.2.0",
-    "prettier": "2.3.2",
-    "typescript": "4.3.5",
+    "prettier": "2.4.1",
+    "typescript": "4.4.4",
     "node-sass": "6.0.1"
   }
 }
