@@ -1,7 +1,7 @@
 import { FC } from 'react';
 import PropTypes from 'prop-types';
-import AddTwoToneIcon from '@material-ui/icons/AddTwoTone';
-import { Typography, Button, Grid } from '@material-ui/core';
+import AddTwoToneIcon from '@mui/icons-material/AddTwoTone';
+import { Typography, Button, Grid } from '@mui/material';
 
 interface PageTitleProps {
     heading?: string;
