@@ -1,6 +1,6 @@
 import { useEffect } from 'react';
 import NProgress from 'nprogress';
-import { Box, CircularProgress } from '@material-ui/core';
+import { Box, CircularProgress } from '@mui/material';
 
 function SuspenseLoader() {
   useEffect(() => {
@@ -13,7 +13,7 @@ function SuspenseLoader() {
 
   return (
     <Box
-      sx={{ width: '100%', height: '100%' }}
+      sx={{ position: 'fixed', left: 0, top: 0, width: '100%', height: '100%' }}
       display="flex"
       alignItems="center"
       justifyContent="center"
