@@ -2,7 +2,7 @@
 import { Helmet } from 'react-helmet-async';
 import Footer from 'src/components/Footer';
 
-import { Container, Grid } from '@material-ui/core';
+import { Grid, Container } from '@mui/material';
 
 import ProfileCover from './ProfileCover';
 import RecentActivity from './RecentActivity';
