@@ -9,7 +9,7 @@ import {
   ListItemText,
   Divider,
   Switch
-} from '@material-ui/core';
+} from '@mui/material';
 
 
 function NotificationsTab() {
