@@ -1,7 +1,7 @@
 import { FC } from 'react';
 import PropTypes from 'prop-types';
 import { Line } from 'react-chartjs-2';
-import { alpha, useTheme } from '@material-ui/core';
+import { alpha, useTheme } from '@mui/material';
 
 interface WatchListColumn1ChartProps {
   data: any[];
