@@ -20,13 +20,13 @@ import {
   DialogTitle,
   Slide,
   Hidden
-} from '@material-ui/core';
-import { experimentalStyled } from '@material-ui/core/styles';
-import { TransitionProps } from '@material-ui/core/transitions';
-import SearchTwoToneIcon from '@material-ui/icons/SearchTwoTone';
-import FindInPageTwoToneIcon from '@material-ui/icons/FindInPageTwoTone';
+} from '@mui/material';
+import { styled } from '@mui/material/styles';
+import { TransitionProps } from '@mui/material/transitions';
+import SearchTwoToneIcon from '@mui/icons-material/SearchTwoTone';
+import FindInPageTwoToneIcon from '@mui/icons-material/FindInPageTwoTone';
 
-import ChevronRightTwoToneIcon from '@material-ui/icons/ChevronRightTwoTone';
+import ChevronRightTwoToneIcon from '@mui/icons-material/ChevronRightTwoTone';
 
 const Transition = forwardRef(function Transition(
   props: TransitionProps & { children?: ReactElement<any, any> },
@@ -35,7 +35,7 @@ const Transition = forwardRef(function Transition(
   return <Slide direction="down" ref={ref} {...props} />;
 });
 
-const DialogWrapper = experimentalStyled(Dialog)(
+const DialogWrapper = styled(Dialog)(
   () => `
     .MuiDialog-container {
         height: auto;
@@ -47,7 +47,7 @@ const DialogWrapper = experimentalStyled(Dialog)(
 `
 );
 
-const SearchInputWrapper = experimentalStyled(TextField)(
+const SearchInputWrapper = styled(TextField)(
   ({ theme }) => `
     background: ${theme.colors.alpha.white[100]};
 
@@ -57,7 +57,7 @@ const SearchInputWrapper = experimentalStyled(TextField)(
 `
 );
 
-const DialogTitleWrapper = experimentalStyled(DialogTitle)(
+const DialogTitleWrapper = styled(DialogTitle)(
   ({ theme }) => `
     background: ${theme.colors.alpha.black[5]};
     padding: ${theme.spacing(3)}
