@@ -1,6 +1,6 @@
 import { FC, ReactNode } from 'react';
-import { experimentalStyled } from '@material-ui/core/styles';
-import { Box } from '@material-ui/core';
+import { styled } from '@mui/material/styles';
+import { Box } from '@mui/material';
 import { Outlet } from 'react-router-dom';
 
 import Sidebar from './Sidebar';
@@ -10,7 +10,7 @@ interface SidebarLayoutProps {
   children?: ReactNode;
 }
 
-const MainWrapper = experimentalStyled(Box)(
+const MainWrapper = styled(Box)(
   ({ theme }) => `
         flex: 1 1 auto;
         display: flex;
@@ -22,7 +22,7 @@ const MainWrapper = experimentalStyled(Box)(
 `
 );
 
-const MainContent = experimentalStyled(Box)(
+const MainContent = styled(Box)(
   ({ theme }) => `
         margin-top: ${theme.header.height};
         flex: 1 1 auto;
